YOLO detection.
